# hospital/api/openapi.json
{
 "components": {
  "schemas": {
...
   "AdmissionType": {
    "enum": [
     "planned",
     "emergency"
    ],
    "title": "AdmissionType",
    "type": "string"
   },
...
   "CaseCreate": {
    "additionalProperties": false,
    "properties": {
     "admission_source": {
      "anyOf": [
       {
        "enum": [
         "ER",
         "OPD",
         "referral"
        ],
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Admission Source"
     },
     "admission_type": {
      "$ref": "#/components/schemas/AdmissionType"
     },
     "admitted_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Admitted At"
     },
     "admitted_on": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Admitted On"
     },
     "claim_type": {
      "$ref": "#/components/schemas/ClaimType"
     },
     "diagnosis_codes": {
      "items": {
       "type": "string"
      },
      "maxItems": 10,
      "title": "Diagnosis Codes",
      "type": "array"
     },
     "discharged_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Discharged At"
     },
     "discharged_on": {
      "anyOf": [
       {
        "format": "date",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Discharged On"
     },
     "patient": {
      "$ref": "#/components/schemas/PatientIn"
     },
     "policy": {
      "$ref": "#/components/schemas/PolicyIn"
     },
     "preauth_ref": {
      "anyOf": [
       {
        "maxLength": 64,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Preauth Ref"
     },
     "procedure_codes": {
      "items": {
       "type": "string"
      },
      "maxItems": 10,
      "title": "Procedure Codes",
      "type": "array"
     },
     "treating_doctor": {
      "anyOf": [
       {
        "maxLength": 120,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Treating Doctor"
     }
    },
    "required": [
     "patient",
     "policy",
     "claim_type",
     "admission_type"
    ],
    "title": "CaseCreate",
    "type": "object"
   },
...
   "ClaimType": {
    "enum": [
     "cashless",
     "reimbursement"
    ],
    "title": "ClaimType",
    "type": "string"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "PatientIn": {
    "additionalProperties": false,
    "properties": {
     "dob": {
      "format": "date",
      "title": "Dob",
      "type": "string"
     },
     "full_name": {
      "maxLength": 120,
      "minLength": 2,
      "title": "Full Name",
      "type": "string"
     },
     "gender": {
      "enum": [
       "M",
       "F",
       "O"
      ],
      "title": "Gender",
      "type": "string"
     },
     "phone": {
      "anyOf": [
       {
        "maxLength": 20,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone"
     },
     "uhid": {
      "maxLength": 40,
      "minLength": 1,
      "title": "Uhid",
      "type": "string"
     }
    },
    "required": [
     "uhid",
     "full_name",
     "dob",
     "gender"
    ],
    "title": "PatientIn",
    "type": "object"
   },
...
   "PolicyIn": {
    "additionalProperties": false,
    "properties": {
     "insurer_name": {
      "maxLength": 64,
      "minLength": 1,
      "title": "Insurer Name",
      "type": "string"
     },
     "member_id": {
      "maxLength": 64,
      "minLength": 1,
      "title": "Member Id",
      "type": "string"
     },
     "policy_number": {
      "maxLength": 64,
      "minLength": 1,
      "title": "Policy Number",
      "type": "string"
     }
    },
    "required": [
     "insurer_name",
     "policy_number",
     "member_id"
    ],
    "title": "PolicyIn",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "Hospital API",
  "version": "1.0.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/v1/cases": {
   "get": {
    "operationId": "listCases",
    "parameters": [
     {
      "in": "query",
      "name": "status",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "items": {
          "type": "string"
         },
         "type": "array"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "in": "query",
      "name": "q",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "in": "query",
      "name": "assigned",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Assigned"
      }
     },
     {
      "in": "query",
      "name": "claim_type",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Claim Type"
      }
     },
     {
      "in": "query",
      "name": "flag",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Flag"
      }
     },
     {
      "in": "query",
      "name": "cursor",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Cursor"
      }
     },
     {
      "in": "query",
      "name": "size",
      "required": false,
      "schema": {
       "default": 25,
       "maximum": 100,
       "minimum": 1,
       "title": "Size",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Listcases",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "List Cases",
    "tags": [
     "cases"
    ]
   },
   "post": {
    "operationId": "createCase",
    "parameters": [
     {
      "in": "query",
      "name": "confirm_patient_update",
      "required": false,
      "schema": {
       "default": false,
       "title": "Confirm Patient Update",
       "type": "boolean"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CaseCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Createcase",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Case",
    "tags": [
     "cases"
    ]
   }
  },
...
 }
}